requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.11.9
pydantic-settings==2.10.1
ultralytics==8.3.202
numpy==2.3.0
opencv-python==4.10.0.84
pillow==11.3.0
python-multipart==0.0.9
dotenv==0.9.9
httpx==0.27.2
